Title: 앞에서부터 삭제하기 2, Time: 844ms, Memory: 34MB, Status: Passed - CodeTree

+++ 250305/앞에서부터 삭제하기 2/delete-it-from-the-beginning-2.java
@@ -18,8 +18,8 @@ public class Main {
             pq.add(now);
             result+=now;
 
-            int cal = result - pq.peek();
-            average = Math.max(cal/(pq.size()-1),average);
+            double cal = result - pq.peek();
+            average = Math.max((double)cal/(pq.size()-1),average);
         }
         DecimalFormat df = new DecimalFormat("0.00");
         System.out.println(df.format(average));
